pyproject.toml
[tool.poetry]
name = "wanikani-mendoukusai-houkoku"
version = "0.1.0"
description = "Generates report of troublesome items in WaniKani"
authors = ["Michael Churilov <michael.churilov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
python-dateutil = "^2.8.1"
DateTime = "^4.3"
pandas = "^1.2.4"
python-dotenv = "^0.17.0"
argparse = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
